split `returns` module
